src/requirements.txt
decorator==3.4.0
Django==1.10.1
django-appconf==1.0.1
django-imagekit==3.2.6
numpy==1.12.0
olefile==0.44
pilkit==1.1.13
Pillow==4.0.0
six==1.10.0
